feat: refactor login and signup forms, removing obsolete components and integrating new ones
