+++ web/src/lib/components/login/index.ts
@@ -1,2 +1,2 @@
-export { default as UserAuthForm } from "./user-auth-form.svelte";
-export { default as UserCreateForm } from "./user-create-form.svelte";
+export { default as LoginForm } from "./login-form.svelte";
+export { default as SignupForm } from "./signup-form.svelte";
